# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EVADIA API",
        "description": "API complète de la plateforme de gestion hôtelière EVADIA. Comprend l'authentification, la gestion admin, le back-office hôtel et les endpoints client.",
        "contact": {
            "name": "EVADIA",
            "email": "contact@evadia.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/hotel/payments": {
            "get": {
                "tags": [
                    "Hôtel - Paiements"
                ],
                "summary": "Liste des paiements",
                "description": "Retourne les paiements reçus pour les réservations de l'hôtel.",
                "operationId": "f63f78bad71659a147272b337d06542c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en_attente",
                                "complete",
                                "echoue",
                                "rembourse"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des paiements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "montant": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 360
                                                    },
                                                    "devise_montant": {
                                                        "type": "string",
                                                        "example": "EUR"
                                                    },
                                                    "statut": {
                                                        "type": "string",
                                                        "example": "complete"
                                                    },
                                                    "transaction_id": {
                                                        "type": "string",
                                                        "example": "txn_abc123"
                                                    },
                                                    "date_paiement": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "reservation": {
                                                        "properties": {
                                                            "code_reservation": {
                                                                "type": "string",
                                                                "example": "EV-a1b2c3d4"
                                                            },
                                                            "client": {
                                                                "properties": {
                                                                    "nom": {
                                                                        "type": "string",
                                                                        "example": "Leclerc"
                                                                    },
                                                                    "prenom": {
                                                                        "type": "string",
                                                                        "example": "Marie"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotel/payments/{id}": {
            "get": {
                "tags": [
                    "Hôtel - Paiements"
                ],
                "summary": "Détails d'un paiement",
                "operationId": "1ef8889a6197caeea3cad18cb84bd848",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du paiement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paiement non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtenu via /api/auth/login ou /api/auth/register",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}